refactor(cli): extract inline test module into dedicated file
Move the inline `#[cfg(test)] mod tests { ... }` block (100 lines)
from `cli_args/add.rs` into `cli_args/add/tests.rs` following
CODE_STYLE_GUIDE.md. No behavioral change.

Part of #358.
